+++ gazon-service/src/main/resources/db/changelog/v-1.1/01-create-bank_card-table.sql
@@ -1,0 +1,6 @@
+CREATE TABLE bank_card (
+    id BIGSERIAL PRIMARY KEY,
+    card_number VARCHAR(255) NOT NULL,
+    due_date DATE NOT NULL,
+    security_code SMALLINT NOT NULL
+);